api show

+++ database/migrations/2023_05_23_151333_create_delivery_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('delivery', function (Blueprint $table) {
-            $table->double('id_delivery');
+            $table->integer('id_delivery');
             $table->string('des_delivery');
             $table->string('time_delivery');
             $table->timestamps();
